chore: move all interfaces into a types folder
