+++ packages/opentelemetry-api-metrics/src/types/BatchObserverResult.ts
