+++ packages/opentelemetry-api-metrics/src/types/BoundInstrument.ts
